Added a way to select a color by hex, you can grab the hsv value through play_pause on main
Color value is HSV not just H like I believe Jason wants so just grab the first value of the set. (H, S, V). H is for hue. That seems to be all Jason needs, but if we need the other values they are there. Just needs to be hooked up through Ardiuno. I believe Jason said in slack that the fisrt value would be Hue then the rest of the 16 values. Or Hue then the percentage of led to go on. Sensitivity still doesnt work, I think it might work through the arduino but not the graph. Also I should prob set the color of the graph to the color that is selected, not sure how since Joey did this, but also might have issue since default color is white and Jason said white isnt allowed so we'll see how that works, also it would conflict with the graph since the back is white.